+++ src/main/java/it/uiip/digitalgarage/roboadvice/persistence/repository/StrategyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StrategyRepository extends JpaRepository<Strategy, Integer> {
+public interface StrategyRepository extends JpaRepository<Strategy, Long> {
 
     List<Strategy> findByUserAndActiveTrue(User user);
